+++ src/settings.js
@@ -248,6 +248,10 @@ var SIMD = 0;
 // Whether closure compiling is being run on this output
 var USE_CLOSURE_COMPILER = 0;
 
+// Enable this flag to have Emscripten treat all Closure warnings as errors,
+// aborting compilation if any warnings were present.
+var CLOSURE_WARNINGS_ARE_ERRORS = 0;
+
 // If set to 1, each asm.js/wasm module export is individually declared with a
 // JavaScript "var" definition. This is the simple and recommended approach.
 // However, this does increase code size (especially if you have many such